Added File command reporter without changing output
