+++ src/robotframework_find_unused/commands/__init__.py
@@ -3,7 +3,7 @@ CLI frontend
 """
 
 from .arguments import ArgumentsOptions, cli_arguments
-from .files import FileOptions, cli_files
+from .files import FileOptions, command_files
 from .keywords import KeywordOptions, cli_keywords
 from .returns import ReturnOptions, cli_returns
 from .variables import VariableOptions, cli_variables
@@ -15,8 +15,8 @@ __all__ = [
     "ReturnOptions",
     "VariableOptions",
     "cli_arguments",
-    "cli_files",
     "cli_keywords",
     "cli_returns",
     "cli_variables",
+    "command_files",
 ]
