+++ src/robotframework_find_unused/commands/files/__init__.py
@@ -1,0 +1,11 @@
+"""
+Find unused files command
+"""
+
+from .files import command_files
+from .options import FileOptions
+
+__all__ = [
+    "FileOptions",
+    "command_files",
+]
